:sparkles: Use of generouted's modal for deleting products with history.state
