+++ src/pages/(dashboard)/products/index.tsx
@@ -1,6 +1,6 @@
+import { useModals } from "@/router";
 import { Button } from "@/components/ui/button";
 import ProductsTable from "./_components/ProductsTable";
-import { useModals } from "@/router";
 
 function ProductsPage() {
   const modals = useModals();
@@ -11,9 +11,7 @@ function ProductsPage() {
         <h1 className="text-2xl font-semibold tracking-tight">All products</h1>
         <Button onClick={() => modals.open("/new-product")}>Create product</Button>
       </div>
-      <div>
-        <ProductsTable />
-      </div>
+      <ProductsTable />
     </>
   );
 }
